requirements.txt
diff-match-patch==20121119
Django==2.0.7
django-import-export==1.0.1
django-postgres-extensions==0.9.3
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.8
psycopg2-binary==2.7.5
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0